requirements.txt
python-telegram-bot[webhooks]==21.4
requests

